requirements.txt
fastapi
uvicorn
nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git
pydub
torchaudio